Added comments and moved AtomicCalculator.py
